+++ src/frontend/screens/Settings/SettingsContext.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+import { SettingsContextType } from 'frontend/types'
+
+const initialContext: SettingsContextType = {
+  getSetting: () => '',
+  setSetting: () => null,
+  config: null,
+  isDefault: true,
+  appName: 'default',
+  runner: 'legendary'
+}
+
+export default React.createContext(initialContext)
