+++ src/frontend/screens/Settings/components/AdvancedSettings/index.tsx
@@ -1,5 +1,7 @@
+import React, { useContext, useEffect, useState } from 'react'
+import { IpcRenderer, Clipboard } from 'electron'
+import { useTranslation } from 'react-i18next'
 import {
-  Backspace,
   CleaningServicesOutlined,
   ContentCopyOutlined,
   DeleteOutline,
@@ -11,17 +13,12 @@ import {
   DeselectOutlined
 } from '@mui/icons-material'
 import classNames from 'classnames'
-import { IpcRenderer, Clipboard } from 'electron'
-import { useTranslation } from 'react-i18next'
-import React, { useContext, useEffect, useState } from 'react'
+import AltLegendaryBin from './AltLegendaryBin'
+import AltGOGdlBin from './AltGOGdlBin'
+import DownloadNoHTTPS from './DownloadNoHTTPS'
+import SettingsContext from '../../SettingsContext'
 import ContextProvider from 'frontend/state/ContextProvider'
-import { AppSettings, GameStatus } from 'common/types'
-import { Path } from 'frontend/types'
-import { ToggleSwitch } from 'frontend/components/UI'
-import { configStore } from 'frontend/helpers/electronStores'
-import TextInputWithIconField from 'frontend/components/UI/TextInputWithIconField'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faFolderOpen } from '@fortawesome/free-solid-svg-icons'
+import { GameStatus } from 'common/types'
 
 interface ElectronProps {
   ipcRenderer: IpcRenderer
@@ -30,27 +27,9 @@ interface ElectronProps {
 
 const { ipcRenderer, clipboard } = window.require('electron') as ElectronProps
 
-interface Props {
-  altLegendaryBin: string
-  altGogdlBin: string
-  downloadNoHttps: boolean
-  setAltLegendaryBin: (value: string) => void
-  setAltGogdlBin: (value: string) => void
-  toggleDownloadNoHttps: () => void
-  settingsToSave: AppSettings
-}
+export default function AdvancedSettings() {
+  const { config } = useContext(SettingsContext)
 
-export const AdvancedSettings = ({
-  altLegendaryBin,
-  altGogdlBin,
-  downloadNoHttps,
-  setAltLegendaryBin,
-  setAltGogdlBin,
-  toggleDownloadNoHttps,
-  settingsToSave
-}: Props) => {
-  const [legendaryVersion, setLegendaryVersion] = useState('')
-  const [gogdlVersion, setGogdlVersion] = useState('')
   const [isCopiedToClipboard, setCopiedToClipboard] = useState(false)
 
   const [eosOverlayInstalled, setEosOverlayInstalled] = useState(false)
@@ -69,11 +48,6 @@ export const AdvancedSettings = ({
   const { t } = useTranslation()
   const isWindows = platform === 'win32'
 
-  const settings = configStore.get('settings') as {
-    altLeg: string
-    altGogdl: string
-  }
-
   useEffect(() => {
     // set copied to clipboard status to true if it's not already set to true
     // used for changing text and color
@@ -85,46 +59,6 @@ export const AdvancedSettings = ({
 
     return () => clearTimeout(timer)
   }, [isCopiedToClipboard])
-
-  useEffect(() => {
-    const getMoreInfo = async () => {
-      configStore.set('settings', {
-        ...settings,
-        altLeg: altLegendaryBin
-      })
-
-      const legendaryVer = await ipcRenderer.invoke('getLegendaryVersion')
-      if (legendaryVer === 'invalid') {
-        setLegendaryVersion('Invalid')
-        setTimeout(() => {
-          setAltLegendaryBin('')
-          return setLegendaryVersion('')
-        }, 3000)
-      }
-      return setLegendaryVersion(legendaryVer)
-    }
-    getMoreInfo()
-  }, [altLegendaryBin])
-
-  useEffect(() => {
-    const getGogdlVersion = async () => {
-      configStore.set('settings', {
-        ...settings,
-        altGogdl: altGogdlBin
-      })
-      const gogdlVersion = await ipcRenderer.invoke('getGogdlVersion')
-      if (gogdlVersion === 'invalid') {
-        setGogdlVersion('Invalid')
-        setTimeout(() => {
-          setAltGogdlBin('')
-          return setGogdlVersion('')
-        }, 3000)
-      }
-      return setGogdlVersion(gogdlVersion)
-    }
-
-    getGogdlVersion()
-  }, [altGogdlBin])
 
   useEffect(() => {
     const getEosStatus = async () => {
@@ -165,32 +99,6 @@ export const AdvancedSettings = ({
     }
     enabledGlobally()
   }, [eosOverlayEnabledGlobally])
-
-  async function handleLegendaryBinary() {
-    return ipcRenderer
-      .invoke('openDialog', {
-        buttonLabel: t('box.choose'),
-        properties: ['openFile'],
-        title: t(
-          'box.choose-legendary-binary',
-          'Select Legendary Binary (needs restart)'
-        )
-      })
-      .then(({ path }: Path) => setAltLegendaryBin(path ? path : ''))
-  }
-
-  async function handleGogdlBinary() {
-    return ipcRenderer
-      .invoke('openDialog', {
-        buttonLabel: t('box.choose'),
-        properties: ['openFile'],
-        title: t(
-          'box.choose-gogdl-binary',
-          'Select GOGDL Binary (needs restart)'
-        )
-      })
-      .then(({ path }: Path) => setAltGogdlBin(path ? path : ''))
-  }
 
   function getMainEosText() {
     if (eosOverlayInstalled && eosOverlayInstallingOrUpdating)
@@ -298,98 +206,14 @@ export const AdvancedSettings = ({
     <div>
       <h3 className="settingSubheader">{t('settings.navbar.advanced')}</h3>
 
-      <TextInputWithIconField
-        htmlId="setting-alt-legendary"
-        label={t(
-          'setting.alt-legendary-bin',
-          'Choose an Alternative Legendary Binary  (needs restart)to use'
-        )}
-        placeholder={t(
-          'placeholder.alt-legendary-bin',
-          'Using built-in Legendary binary...'
-        )}
-        value={altLegendaryBin.replaceAll("'", '')}
-        onChange={(event) => setAltLegendaryBin(event.target.value)}
-        icon={
-          !altLegendaryBin.length ? (
-            <FontAwesomeIcon
-              icon={faFolderOpen}
-              data-testid="setLegendaryBinaryButton"
-              style={{
-                color: altLegendaryBin.length ? 'transparent' : 'currentColor'
-              }}
-            />
-          ) : (
-            <Backspace
-              data-testid="setLegendaryBinaryBackspace"
-              style={{ color: 'currentColor' }}
-            />
-          )
-        }
-        onIconClick={
-          !altLegendaryBin.length
-            ? async () => handleLegendaryBinary()
-            : () => setAltLegendaryBin('')
-        }
-        afterInput={
-          <span className="smallMessage">
-            {t('other.legendary-version', 'Legendary Version: ')}
-            {legendaryVersion}
-          </span>
-        }
-      />
+      <AltLegendaryBin />
 
-      <TextInputWithIconField
-        label={t(
-          'setting.alt-gogdl-bin',
-          'Choose an Alternative GOGDL Binary to use (needs restart)'
-        )}
-        htmlId="setting-alt-gogdl"
-        placeholder={t(
-          'placeholder.alt-gogdl-bin',
-          'Using built-in GOGDL binary...'
-        )}
-        value={altGogdlBin.replaceAll("'", '')}
-        onChange={(event) => setAltGogdlBin(event.target.value)}
-        icon={
-          !altGogdlBin.length ? (
-            <FontAwesomeIcon
-              icon={faFolderOpen}
-              data-testid="setGogdlBinaryButton"
-              style={{
-                color: altGogdlBin.length ? 'transparent' : 'currentColor'
-              }}
-            />
-          ) : (
-            <Backspace
-              data-testid="setGogdlBinaryBackspace"
-              style={{ color: '#currentColor' }}
-            />
-          )
-        }
-        onIconClick={
-          !altGogdlBin.length
-            ? async () => handleGogdlBinary()
-            : () => setAltGogdlBin('')
-        }
-        afterInput={
-          <span className="smallMessage">
-            {t('other.gogdl-version', 'GOGDL Version: ')}
-            {gogdlVersion}
-          </span>
-        }
-      />
+      <AltGOGdlBin />
 
-      <ToggleSwitch
-        htmlId="downloadNoHttps"
-        value={downloadNoHttps}
-        handleChange={toggleDownloadNoHttps}
-        title={t(
-          'setting.download-no-https',
-          'Download games without HTTPS (useful for CDNs e.g. LanCache)'
-        )}
-      />
+      <DownloadNoHTTPS />
+
       <hr />
+
       <div className="eosSettings">
         <h3>EOS Overlay</h3>
         <div>{getMainEosText()}</div>
@@ -510,7 +334,7 @@ export const AdvancedSettings = ({
             isSuccess: isCopiedToClipboard
           })}
           onClick={() => {
-            clipboard.writeText(JSON.stringify({ ...settingsToSave }, null, 2))
+            clipboard.writeText(JSON.stringify({ ...config }, null, 2))
             setCopiedToClipboard(true)
           }}
         >
