+++ src/frontend/screens/Settings/components/LogSettings/index.tsx
@@ -1,12 +1,12 @@
+import React, { useContext, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
 import { faFolderOpen } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import ContentCopyIcon from '@mui/icons-material/ContentCopy'
-import React, { useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
 import { UpdateComponent } from 'frontend/components/UI'
-import './index.css'
-
 import { ipcRenderer } from 'frontend/helpers'
+import SettingsContext from '../../SettingsContext'
+import './index.css'
 
 interface LogBoxProps {
   logFileContent: string
@@ -59,17 +59,13 @@ const LogBox: React.FC<LogBoxProps> = ({ logFileContent }) => {
   )
 }
 
-export interface LogSettingsProps {
-  isDefault: boolean
-  appName: string
-}
-
-export default function LogSettings({ isDefault, appName }: LogSettingsProps) {
+export default function LogSettings() {
   const { t } = useTranslation()
   const [logFileContent, setLogFileContent] = useState<string>('')
   const [logFileExist, setLogFileExist] = useState<boolean>(false)
   const [defaultLast, setDefaultLast] = useState<boolean>(false)
   const [refreshing, setRefreshing] = useState<boolean>(true)
+  const { appName, isDefault } = useContext(SettingsContext)
 
   const getLogContent = () => {
     ipcRenderer
