+++ src/frontend/screens/Settings/components/OtherSettings/PrimeRun.tsx
@@ -1,0 +1,40 @@
+import React, { useContext } from 'react'
+import { useTranslation } from 'react-i18next'
+import { ToggleSwitch } from 'frontend/components/UI'
+import useSetting from 'frontend/hooks/useSetting'
+import ContextProvider from 'frontend/state/ContextProvider'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faCircleInfo } from '@fortawesome/free-solid-svg-icons'
+
+const PrimeRun = () => {
+  const { t } = useTranslation()
+  const { platform } = useContext(ContextProvider)
+  const isLinux = platform === 'linux'
+  const [primeRun, setPrimeRun] = useSetting<boolean>('primeRun', false)
+
+  if (!isLinux) {
+    return <></>
+  }
+
+  return (
+    <div className="toggleRow">
+      <ToggleSwitch
+        htmlId="primerun"
+        value={primeRun}
+        handleChange={() => setPrimeRun(!primeRun)}
+        title={t('setting.primerun', 'Use Dedicated Graphics Card')}
+      />
+
+      <FontAwesomeIcon
+        className="helpIcon"
+        icon={faCircleInfo}
+        title={t(
+          'help.primerun',
+          'Use dedicated graphics card to render game on multi-GPU systems. Only needed on gaming laptops or desktops that use a headless GPU for rendering (NVIDIA Optimus, AMD CrossFire)'
+        )}
+      />
+    </div>
+  )
+}
+
+export default PrimeRun
