+++ src/frontend/screens/Settings/components/WineExtensions/index.tsx
@@ -1,106 +1,12 @@
-import { faCircleInfo, faSyncAlt } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React, { useState } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
-import { ToggleSwitch } from 'frontend/components/UI'
-import { ipcRenderer } from 'frontend/helpers'
-import { WineInstallation } from 'common/types'
+import BattlEyeRuntime from './BattlEyeRuntime'
+import EacRuntime from './EacRuntime'
+import AutoDXVK from './AutoDXVK'
+import AutoVKD3D from './AutoVKD3D'
 
-interface Props {
-  wineVersion: WineInstallation
-  winePrefix: string
-  eacRuntime: boolean
-  toggleEacRuntime: () => void
-  gameMode: boolean
-  toggleGameMode: () => void
-  battlEyeRuntime: boolean
-  toggleBattlEyeRuntime: () => void
-  autoInstallDxvk: boolean
-  toggleAutoInstallDxvk: () => void
-  autoInstallVkd3d: boolean
-  toggleAutoInstallVkd3d: () => void
-}
-
-export default function WineExtensions({
-  wineVersion,
-  winePrefix,
-  eacRuntime,
-  gameMode,
-  battlEyeRuntime,
-  autoInstallDxvk,
-  autoInstallVkd3d,
-  toggleEacRuntime,
-  toggleGameMode,
-  toggleBattlEyeRuntime,
-  toggleAutoInstallDxvk,
-  toggleAutoInstallVkd3d
-}: Props) {
-  const [eacInstalling, setEacInstalling] = useState(false)
-  const [battlEyeInstalling, setBattlEyeInstalling] = useState(false)
-  const isProton = wineVersion.type === 'proton'
+export default function WineExtensions() {
   const { t } = useTranslation()
-
-  const handleEacRuntime = async () => {
-    if (!eacRuntime) {
-      if (!gameMode) {
-        const isFlatpak = await ipcRenderer.invoke('isFlatpak')
-        if (isFlatpak) {
-          const { response } = await ipcRenderer.invoke('openMessageBox', {
-            message: t(
-              'settings.eacRuntime.gameModeRequired.message',
-              'GameMode is required for the EAC runtime to work on Flatpak. Do you want to enable it now?'
-            ),
-            title: t(
-              'settings.eacRuntime.gameModeRequired.title',
-              'GameMode required'
-            ),
-            buttons: [t('box.yes'), t('box.no')]
-          })
-          if (response === 1) {
-            return
-          }
-          toggleGameMode()
-        }
-      }
-      const isInstalled = await ipcRenderer.invoke(
-        'isRuntimeInstalled',
-        'eac_runtime'
-      )
-      if (!isInstalled) {
-        setEacInstalling(true)
-        const success = await ipcRenderer.invoke(
-          'downloadRuntime',
-          'eac_runtime'
-        )
-        setEacInstalling(false)
-        if (!success) {
-          return
-        }
-      }
-    }
-    toggleEacRuntime()
-  }
-
-  const handleBattlEyeRuntime = async () => {
-    if (!battlEyeRuntime) {
-      const isInstalled = await ipcRenderer.invoke(
-        'isRuntimeInstalled',
-        'battleye_runtime'
-      )
-      if (!isInstalled) {
-        setBattlEyeInstalling(true)
-        const success = await ipcRenderer.invoke(
-          'downloadRuntime',
-          'battleye_runtime'
-        )
-        setBattlEyeInstalling(false)
-        if (!success) {
-          return
-        }
-      }
-    }
-    toggleBattlEyeRuntime()
-  }
 
   return (
     <>
@@ -108,92 +14,13 @@ export default function WineExtensions({
         {t('settings.navbar.wineExt', 'Wine Extensions')}
       </h3>
 
-      {!isProton && (
-        <>
-          <div className="toggleRow">
-            <ToggleSwitch
-              htmlId="autodxvk"
-              value={autoInstallDxvk}
-              handleChange={() => {
-                const action = autoInstallDxvk ? 'restore' : 'backup'
-                ipcRenderer.send('toggleDXVK', [
-                  { winePrefix, winePath: wineVersion.bin },
-                  action
-                ])
-                return toggleAutoInstallDxvk()
-              }}
-              title={t(
-                'setting.autodxvk',
-                'Auto Install/Update DXVK on Prefix'
-              )}
-            />
+      <AutoDXVK />
 
-            <FontAwesomeIcon
-              className="helpIcon"
-              icon={faCircleInfo}
-              title={t(
-                'help.dxvk',
-                'DXVK is a Vulkan-based translational layer for DirectX 9, 10 and 11 games. Enabling may improve compatibility. Might cause issues especially for older DirectX games.'
-              )}
-            />
-          </div>
-          <div className="toggleRow">
-            <ToggleSwitch
-              htmlId="autovkd3d"
-              value={autoInstallVkd3d}
-              handleChange={() => {
-                const action = autoInstallVkd3d ? 'restore' : 'backup'
-                ipcRenderer.send('toggleVKD3D', [
-                  { winePrefix, winePath: wineVersion.bin },
-                  action
-                ])
-                return toggleAutoInstallVkd3d()
-              }}
-              title={t(
-                'setting.autovkd3d',
-                'Auto Install/Update VKD3D on Prefix'
-              )}
-            />
+      <AutoVKD3D />
 
-            <FontAwesomeIcon
-              className="helpIcon"
-              icon={faCircleInfo}
-              title={t(
-                'help.vkd3d',
-                'VKD3D is a Vulkan-based translational layer for DirectX 12 games. Enabling may improve compatibility significantly. Has no effect on older DirectX games.'
-              )}
-            />
-          </div>
-        </>
-      )}
+      <EacRuntime />
 
-      <ToggleSwitch
-        htmlId="eacRuntime"
-        value={eacRuntime}
-        handleChange={handleEacRuntime}
-        title={t('settings.eacRuntime.name', 'EasyAntiCheat Runtime')}
-      />
-      {eacInstalling && (
-        <span>
-          <FontAwesomeIcon className="fa-spin" icon={faSyncAlt} />
-          {t('settings.eacRuntime.installing', 'Installing EAC Runtime...')}
-        </span>
-      )}
-      <ToggleSwitch
-        htmlId="battlEyeRuntime"
-        value={battlEyeRuntime}
-        handleChange={handleBattlEyeRuntime}
-        title={t('settings.battlEyeRuntime.name', 'BattlEye AntiCheat Runtime')}
-      />
-      {battlEyeInstalling && (
-        <span>
-          <FontAwesomeIcon className="fa-spin" icon={faSyncAlt} />
-          {t(
-            'settings.battlEyeRuntime.installing',
-            'Installing BattlEye Runtime...'
-          )}
-        </span>
-      )}
+      <BattlEyeRuntime />
     </>
   )
 }
