+++ src/frontend/screens/Settings/components/WineSettings/WineVersionSelector.tsx
@@ -1,0 +1,82 @@
+import React, { useContext, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { InfoBox, SelectField } from 'frontend/components/UI'
+import { ipcRenderer } from 'frontend/helpers'
+import ContextProvider from 'frontend/state/ContextProvider'
+import { WineInstallation } from 'common/types'
+import useSetting from 'frontend/hooks/useSetting'
+import { defaultWineVersion } from '.'
+
+export default function WineVersionSelector() {
+  const { t } = useTranslation()
+  const { platform } = useContext(ContextProvider)
+  const isLinux = platform === 'linux'
+
+  const [wineVersion, setWineVersion] = useSetting<WineInstallation>(
+    'wineVersion',
+    defaultWineVersion
+  )
+  const [altWine, setAltWine] = useState<WineInstallation[]>([])
+
+  useEffect(() => {
+    const getAltWine = async () => {
+      const wineList: WineInstallation[] = await ipcRenderer.invoke(
+        'getAlternativeWine'
+      )
+      setAltWine(wineList)
+      // Avoids not updating wine config when having one wine install only
+      if (wineList && wineList.length === 1) {
+        setWineVersion(wineList[0])
+      }
+    }
+    getAltWine()
+  }, [])
+
+  return (
+    <SelectField
+      label={
+        isLinux
+          ? t('setting.wineversion')
+          : t('setting.crossover-version', 'Crossover Version')
+      }
+      htmlId="setWineVersion"
+      onChange={(event) =>
+        setWineVersion(
+          altWine.filter(({ name }) => name === event.target.value)[0]
+        )
+      }
+      value={wineVersion.name}
+      afterSelect={
+        <>
+          {isLinux && (
+            <InfoBox text={t('infobox.wine-path', 'Wine Path')}>
+              {wineVersion.bin}
+            </InfoBox>
+          )}
+          {isLinux && (
+            <InfoBox text="infobox.help">
+              <span>{t('help.wine.part1')}</span>
+              <ul>
+                <i>
+                  <li>~/.config/heroic/tools/wine</li>
+                  <li>~/.config/heroic/tools/proton</li>
+                  <li>~/.steam/root/compatibilitytools.d</li>
+                  <li>~/.steam/steamapps/common</li>
+                  <li>~/.local/share/lutris/runners/wine</li>
+                  <li>~/.var/app/com.valvesoftware.Steam (Steam Flatpak)</li>
+                  <li>/usr/share/steam</li>
+                  <li>Everywhere on the system (CrossOver Mac)</li>
+                </i>
+              </ul>
+              <span>{t('help.wine.part2')}</span>
+            </InfoBox>
+          )}
+        </>
+      }
+    >
+      {altWine.map(({ name }) => (
+        <option key={name}>{name}</option>
+      ))}
+    </SelectField>
+  )
+}
