+++ src/frontend/screens/Settings/components/index.tsx
@@ -1,6 +1,7 @@
 export { default as GeneralSettings } from './GeneralSettings'
 export { default as OtherSettings } from './OtherSettings'
-export { LegendarySyncSaves, GOGSyncSaves } from './SyncSaves'
 export { default as Tools } from './Tools'
 export { default as WineSettings } from './WineSettings'
 export { default as WineExtensions } from './WineExtensions'
+export { default as AdvancedSettings } from './AdvancedSettings'
+export { default as SyncSaves } from './SyncSaves'
